// apps/api-gateway/src/shared/common.enum.ts
export enum OptionType {
  CALL = 'CE',
  PUT = 'PE',
}

export enum InstrumentType {
  STOCK = 'STOCK',
  FUTURE = 'FUTURE',
  OPTION = 'OPTION',
  EQUITY = 'EQUITY',
  INDEX = 'INDEX',
}

export enum Timeframe {
  ONE_MINUTE = '1m',
  THREE_MINUTES = '3m',
  FIVE_MINUTES = '5m',
  FIFTEEN_MINUTES = '15m',
  ONE_HOUR = '1h',
  ONE_DAY = '1d',
}